Amended filter layout and styling
