--- src/views/Results/Filters/CategoryFilter/CategoryFilter.scss
@@ -1,20 +1,0 @@
-@import '../../../../styles/mixins.scss';
-@import '../../../../styles/colors.scss';
-@import '../../../../styles/fonts.scss';
-@import '../../../../styles/breakpoints.scss';
-
-.category__filters {
-  @include for-tablet-landscape-up {
-    justify-content: flex-end !important;
-    border-left: 3px solid #d8d8d8;
-  }
-
-  @include for-tablet-portrait-down {
-    padding: 0 !important;
-  }
-}
-
-.category_filters--header--cost {
-  @include heading-m;
-  margin-bottom: space(16);
-}
