--- src/views/Results/Filters/CategoryFilter/index.ts
@@ -1,3 +1,0 @@
-import CategoryFilter from './CategoryFilter';
-
-export default CategoryFilter;
