--- src/views/Results/Filters/KeywordFilter/index.ts
@@ -1,3 +1,0 @@
-import KeywordFilter from './KeywordFilter';
-
-export default KeywordFilter;
